Replace Driver with App.java
